fix for python3

+++ gadgets/python/gadgets/remove_2x_oversampling.py
@@ -11,7 +11,7 @@ class Remove2xOversampling(Gadget):
 
     def process(self, acq, data):
         orig_size = list(data.shape);
-        data2 = data.reshape([data.shape[0],(data.size/data.shape[0])])
+        data2 = data.reshape([data.shape[0], int(data.size/data.shape[0])])
         new_length = data2.shape[0]>>1
         data2 = transform.transform_image_to_kspace(transform.transform_kspace_to_image(data2,dim=(0,))[(0+(new_length>>1)):(new_length+(new_length>>1)),:],dim=(0,))
         orig_size[0] = new_length
